Handle includes in testcases while generating the testcases.

+++ carbon/test/generate-testcases.sh
@@ -3,34 +3,58 @@
 succes_cases=$( for i in $( ls success/*.test ); do echo $( basename $i ); done )
 fail_cases=$( for i in $( ls fail/*.test ); do echo $( basename $i ); done )
 
+DEPENDENCIES=
+function generate_dependencies {
+    local TESTCASE=$1
+    local TYPE=$2
+    local DEPS=$( grep -E "^include" ${TESTCASE} | sort -u | cut -d' ' -f2 )
+    DEPENDENCIES=""
+    for DEP in ${DEPS}
+    do
+	DEPENDENCIES="${DEPENDENCIES} ${TYPE}/${DEP}.co2"
+    done
+}
+
 function generate_testcase {
     local TYPE=$1 # success/fail
     local NAME=$2
 
     local BASENAME=$( basename ${NAME} .test )
 
-    local TARGET=${TYPE}_${BASENAME}.sh
+    local TARGET=${TYPE}/${BASENAME}.sh
 
-    echo "#!/bin/bash -u" > ${TARGET}
-    echo >> ${TARGET}
+    GENERATE=$( grep ${TARGET} testdonotcreate | wc -l )
 
-    if [[ "${TYPE}" == "success" ]];
+
+    if [[ ! -e ${TARGET} || ${GENERATE} != 1 ]];
     then
-	echo "./run_succes_test.sh ${TYPE}/${BASENAME}.test" >> ${TARGET}
-    else
-	echo "./run_fail_test.sh ${TYPE}/${BASENAME}.test" >> ${TARGET}
+	echo "#!/bin/bash -u" > ${TARGET}
+	echo >> ${TARGET}
+
+	if [[ "${TYPE}" == "success" ]];
+	then
+	    generate_dependencies ${TYPE}/${BASENAME}.test ${TYPE}
+	    echo "./run_succes_test.sh ${DEPENDENCIES} ${TYPE}/${BASENAME}.test" >> ${TARGET}
+	else
+	    generate_dependencies ${TYPE}/${BASENAME}.test ${TYPE}
+	    echo "./run_fail_test.sh ${DEPENDENCIES} ${TYPE}/${BASENAME}.test" >> ${TARGET}
+	fi
+
+	chmod +x ${TARGET}
     fi
 
-    chmod +x ${TARGET}
-
-    echo '\' >> Makefile.am
-    echo -n ${TARGET} >> Makefile.am
+    IGNORE=$( grep ${TARGET} testignore | wc -l )
+    if [[ ${IGNORE} == 0 ]];
+    then
+	echo '\' >> Makefile.am
+	echo -n ${TARGET} >> Makefile.am
+    fi
 }
 
 function generate_testcases {
     local TYPE=$1
     local CASES=$2
-    
+    local TESTCASE=
     for TESTCASE in ${CASES}
     do
 	generate_testcase ${TYPE} ${TESTCASE}
@@ -40,8 +64,6 @@ function generate_testcases {
 function generate_makefile_am_header {
     echo 'EXTRA_DIST = success fail generate-testcases.sh run_succes_test.sh run_fail_test.sh'
     echo 'MAINTAINERCLEANFILES = Makefile.in'
-    echo
-    echo '#SUBDIRS = success fail'
     echo
     echo 'targetdir = target'
     echo
@@ -53,7 +75,7 @@ function generate_makefile_am_header {
     echo 'clean-local:'
     echo '	-rm -f ${TESTS}'
     echo
-    echo '${TESTS}: ./generate-testcases.sh'
+    echo '${TESTS}, Makefile.am: ./generate-testcases.sh'
     echo '	./generate-testcases.sh'
     echo
     echo -n 'TESTS='
